+++ src/pretix/base/models/event.py
@@ -64,6 +64,7 @@ from django.utils.translation import gettext, gettext_lazy as _
 from django_scopes import ScopedManager, scopes_disabled
 from i18nfield.fields import I18nCharField, I18nTextField
 
+from pretix.base.middleware.timemachine import time_machine_now
 from pretix.base.models.base import LoggedModel
 from pretix.base.models.fields import MultiStringField
 from pretix.base.reldate import RelativeDateWrapper
@@ -76,7 +77,6 @@ from pretix.helpers.thumb import get_thumbnail
 
 from ..settings import settings_hierarkey
 from .organizer import Organizer, Team
-from ...presale.timemachine import time_machine_now
 
 logger = logging.getLogger(__name__)
