+++ src/pretix/base/services/memberships.py
@@ -25,15 +25,14 @@ from typing import List, Optional
 from dateutil.relativedelta import relativedelta
 from django.core.exceptions import ValidationError
 from django.utils.formats import date_format
-from django.utils.timezone import now
 from django.utils.translation import gettext_lazy as _
 
+from pretix.base.middleware.timemachine import time_machine_now
 from pretix.base.models import (
     AbstractPosition, CartPosition, Customer, Event, Item, Membership, Order,
     OrderPosition, SubEvent,
 )
 from pretix.helpers import OF_SELF
-from pretix.presale.timemachine import time_machine_now
 
 
 def membership_validity(item: Item, subevent: Optional[SubEvent], event: Event):
